Extract Slot into its own module
