+++ wallet-api/src/Wallet/Emulator/Types.hs
@@ -92,7 +92,7 @@ import           Ledger                     (Address, Block, Blockchain, Slot, T
                                              TxOutRef, Value, hashTx, lastSlot, pubKeyAddress, pubKeyTxIn, pubKeyTxOut,
                                              txOutAddress)
 import qualified Ledger.Index               as Index
-import qualified Ledger.Interval            as Interval
+import qualified Ledger.Slot                as Slot
 import qualified Ledger.Value               as Value
 import           Wallet.API                 (EventHandler (..), EventTrigger, KeyPair (..), WalletAPI (..),
                                              WalletAPIError (..), WalletDiagnostics (..), WalletLog (..), addresses,
@@ -480,7 +480,7 @@ validateBlock currentSlot idx txns =
 
 -- | Check whether the given transaction can be validated in the given slot.
 canValidateNow :: Slot -> Tx -> Bool
-canValidateNow currentSlot tx = $$(Interval.member) currentSlot (txValidRange tx)
+canValidateNow currentSlot tx = $$(Slot.member) currentSlot (txValidRange tx)
 
 mkEvent :: Tx -> Maybe Index.ValidationError -> EmulatorEvent
 mkEvent t result =
