+++ wallet-api/wallet-api.cabal
@@ -2,7 +2,7 @@ cabal-version: 2.2
 name: wallet-api
 version: 0.1.0.0
 license: Apache-2.0
-license-files: 
+license-files:
   LICENSE
   NOTICE
 maintainer: michael.peyton-jones@iohk.io
@@ -59,6 +59,7 @@ library
         Ledger.Map.TH
         Ledger.These
         Ledger.These.TH
+        Ledger.Slot
         Ledger.Types
         Ledger.Validation
         Ledger.Index
